+++ WebApplication/src/main-inbox.js
@@ -32,77 +32,92 @@ axios.get('../api/pre-login-configuration').then((config) => {
         return;
     }
 
-    let keycloackConfig = null;
+    // check if a token is passed in the query args (in case we open the inbox with an inbox-link)
+    const params = new URLSearchParams(window.location.search);
+    const tokenValue = params.get("token");
+    if (tokenValue)
+    {
+        localStorage.setItem("token", tokenValue);
+        orthancApi.updateAuthHeader("token");
 
-    if (config.data['Keycloak'] && config.data['Keycloak'] != null && config.data['Keycloak']['Enable']) {
-        if (configData['Inbox']['EnableAnonymousAccess']) {
-
-            console.log("Keycloak is enabled but AnonymousAccess to Inbox is enabled -> all users are considered as anonymous");
-            app.mount('#app-inbox');
-
-        } else {
-            console.log("Keycloak is enabled");
-
-            keycloackConfig = config.data['Keycloak']
-
-            let initOptions = {
-                url: keycloackConfig['Url'],
-                realm: keycloackConfig['Realm'],
-                clientId: keycloackConfig['ClientId'],
-                onLoad: 'login-required'
-            }
-
-            window.keycloak = new Keycloak(initOptions);
-
-            window.keycloak.init({ onLoad: initOptions.onLoad }).then(async (auth) => {
-
-                if (!auth) {
-                    window.location.reload();
-                } else {
-                    console.log("Authenticated");
-                }
-
-                localStorage.setItem("vue-token", window.keycloak.token);
-                localStorage.setItem("vue-refresh-token", window.keycloak.refreshToken);
-                orthancApi.updateAuthHeader();
-
-                app.mount('#app-inbox');
-                console.log("App mounted with keycloak, current route is ", router.currentRoute.value.fullPath);
-
-                // keycloak includes state, code and session_state -> the router does not like them -> remove them
-                const params = new URLSearchParams(router.currentRoute.value.fullPath);
-                params.delete('state');
-                params.delete('code');
-                params.delete('session_state');
-                const cleanedRoute = decodeURIComponent(params.toString()).replace('/=', '/');
-                console.log("App mounted, moving to cleaned route ", cleanedRoute);
-                await router.push(cleanedRoute);
-
-                // programm token refresh at regular interval
-                setInterval(() => {
-                    window.keycloak.updateToken(70).then((refreshed) => {
-                        if (refreshed) {
-                            console.log('Token refreshed');
-                            localStorage.setItem("vue-token", window.keycloak.token);
-                            localStorage.setItem("vue-refresh-token", window.keycloak.refreshToken);
-                            orthancApi.updateAuthHeader();
-                        } else {
-                            console.log('Token not refreshed, still valid for '
-                                + Math.round(window.keycloak.tokenParsed.exp + window.keycloak.timeSkew - new Date().getTime() / 1000) + ' seconds');
-                        }
-                    }).catch(() => {
-                        console.log('Failed to refresh token');
-                    });
-
-                }, 60000)
-
-            }).catch(() => {
-                console.log("Could not connect to Keycloak");
-            });
-        }
-    } else {
-        console.log("Keycloak is disabled");
+        console.log("Opening the inbox with an inbox-link");
 
         app.mount('#app-inbox')
+    }
+    else
+    {
+        let keycloackConfig = null;
+
+        if (config.data['Keycloak'] && config.data['Keycloak'] != null && config.data['Keycloak']['Enable']) {
+            if (configData['Inbox']['EnableAnonymousAccess']) {
+
+                console.log("Keycloak is enabled but AnonymousAccess to Inbox is enabled -> all users are considered as anonymous");
+                app.mount('#app-inbox');
+
+            } else {
+                console.log("Keycloak is enabled");
+
+                keycloackConfig = config.data['Keycloak']
+
+                let initOptions = {
+                    url: keycloackConfig['Url'],
+                    realm: keycloackConfig['Realm'],
+                    clientId: keycloackConfig['ClientId'],
+                    onLoad: 'login-required'
+                }
+
+                window.keycloak = new Keycloak(initOptions);
+
+                window.keycloak.init({ onLoad: initOptions.onLoad }).then(async (auth) => {
+
+                    if (!auth) {
+                        window.location.reload();
+                    } else {
+                        console.log("Authenticated");
+                    }
+
+                    localStorage.setItem("vue-token", window.keycloak.token);
+                    localStorage.setItem("vue-refresh-token", window.keycloak.refreshToken);
+                    orthancApi.updateAuthHeader();
+
+                    app.mount('#app-inbox');
+                    console.log("App mounted with keycloak, current route is ", router.currentRoute.value.fullPath);
+
+                    // keycloak includes state, code and session_state -> the router does not like them -> remove them
+                    const params = new URLSearchParams(router.currentRoute.value.fullPath);
+                    params.delete('state');
+                    params.delete('code');
+                    params.delete('session_state');
+                    const cleanedRoute = decodeURIComponent(params.toString()).replace('/=', '/');
+                    console.log("App mounted, moving to cleaned route ", cleanedRoute);
+                    await router.push(cleanedRoute);
+
+                    // programm token refresh at regular interval
+                    setInterval(() => {
+                        window.keycloak.updateToken(70).then((refreshed) => {
+                            if (refreshed) {
+                                console.log('Token refreshed');
+                                localStorage.setItem("vue-token", window.keycloak.token);
+                                localStorage.setItem("vue-refresh-token", window.keycloak.refreshToken);
+                                orthancApi.updateAuthHeader();
+                            } else {
+                                console.log('Token not refreshed, still valid for '
+                                    + Math.round(window.keycloak.tokenParsed.exp + window.keycloak.timeSkew - new Date().getTime() / 1000) + ' seconds');
+                            }
+                        }).catch(() => {
+                            console.log('Failed to refresh token');
+                        });
+
+                    }, 60000)
+
+                }).catch(() => {
+                    console.log("Could not connect to Keycloak");
+                });
+            }
+        } else {
+            console.log("Keycloak is disabled");
+
+            app.mount('#app-inbox')
+        }
     }
 });